add screens to example app
