+++ example/android/app/src/main/kotlin/com/klarna/inapp/sdk/flutter_klarna_inapp_sdk_example/MainActivity.kt
@@ -1,12 +1,11 @@
 package com.klarna.inapp.sdk.flutter_klarna_inapp_sdk_example
 
-import androidx.annotation.NonNull;
 import io.flutter.embedding.android.FlutterActivity
 import io.flutter.embedding.engine.FlutterEngine
 import io.flutter.plugins.GeneratedPluginRegistrant
 
-class MainActivity: FlutterActivity() {
-    override fun configureFlutterEngine(@NonNull flutterEngine: FlutterEngine) {
+class MainActivity : FlutterActivity() {
+    override fun configureFlutterEngine(flutterEngine: FlutterEngine) {
         GeneratedPluginRegistrant.registerWith(flutterEngine);
     }
 }
